+++ desktop/data_modules/unMainDataModule.pas
@@ -26,6 +26,7 @@ type
     function ValidatedMail(Hash: string): boolean;
     function ResendValidateMail(Hash: string): boolean;
     procedure ProcessMigrations;
+    procedure UpdateMigrationVersion(Version: Integer);
   end;
 
 var
@@ -98,15 +99,117 @@ begin
       Query.SQL.Add('   id INTEGER PRIMARY KEY AUTOINCREMENT,  ');
       Query.SQL.Add('   code VARCHAR(100) NOT NULL);           ');
       Query.ExecSQL;
-      Query.Close; 
+      Query.Close;
       
-      Query.SQL.Clear;    
-      Query.Params.Clear;
-      Query.SQL.Add('UPDATE parameter SET param_value = :param_value WHERE param_name = :param_name;');
-      Query.Params.ParamByName('param_name').Value := PARAM_VERSION_NAME;
-      Query.Params.ParamByName('param_value').Value := IntToStr(Version);
+      UpdateMigrationVersion(Version);
+    end;
+
+    // Populate Project Type
+    if Version <= 1 then
+    begin
+      Version := Version + 1;
+
+      Query.SQL.Clear;
+      Query.SQL.Add('INSERT INTO project_type (code) VALUES ');
+      Query.SQL.Add('   (:project_type_candidate),          ');
+      Query.SQL.Add('   (:project_type_recruiter)           ');
+      Query.Params.ParamByName('project_type_candidate').AsString := 'PROJECT_TYPE_CANDIDATE';
+      Query.Params.ParamByName('project_type_recruiter').AsString := 'PROJECT_TYPE_RECRUITER';
       Query.ExecSQL;
       Query.Close;
+
+      UpdateMigrationVersion(Version);
+    end;
+
+    // Create Project Table
+    if Version <= 2 then
+    begin
+      Version := Version + 1;
+
+      Query.SQL.Clear;
+      Query.SQL.Add('CREATE TABLE IF NOT EXISTS project (                   ');
+      Query.SQL.Add('   id INTEGER PRIMARY KEY AUTOINCREMENT,               ');
+      Query.SQL.Add('   name VARCHAR(100) NOT NULL,                         ');
+      Query.SQL.Add('   type_id INTEGER NOT NULL,                           ');
+      Query.SQL.Add('   description TEXT,                                   ');
+      Query.SQL.Add('   FOREIGN KEY (type_id) REFERENCES project_type(id)); ');
+      Query.ExecSQL;
+      Query.Close;
+
+      UpdateMigrationVersion(Version);
+    end;
+
+    // Create Folder Table
+    if Version <= 3 then
+    begin
+      Version := Version + 1;
+
+      Query.SQL.Clear;
+      Query.SQL.Add('CREATE TABLE IF NOT EXISTS folder (                  ');
+      Query.SQL.Add('   id INTEGER PRIMARY KEY AUTOINCREMENT,             ');
+      Query.SQL.Add('   path VARCHAR(500) NOT NULL,                       ');
+      Query.SQL.Add('   project_id INTEGER NOT NULL,                      ');
+      Query.SQL.Add('   FOREIGN KEY (project_id) REFERENCES project(id)); ');
+      Query.ExecSQL;
+      Query.Close;
+
+      UpdateMigrationVersion(Version);
+    end;
+
+    // Create File Table
+    if Version <= 4 then
+    begin
+      Version := Version + 1;
+
+      Query.SQL.Clear;
+      Query.SQL.Add('CREATE TABLE IF NOT EXISTS file (                  ');
+      Query.SQL.Add('   id INTEGER PRIMARY KEY AUTOINCREMENT,           ');
+      Query.SQL.Add('   filename VARCHAR(100) NOT NULL,                 ');
+      Query.SQL.Add('   candidate VARCHAR(200),                         ');
+      Query.SQL.Add('   remote_filename VARCHAR(100),                   ');
+      Query.SQL.Add('   folder_id INTEGER NOT NULL,                     ');
+      Query.SQL.Add('   FOREIGN KEY (folder_id) REFERENCES folder(id)); ');
+      Query.ExecSQL;
+      Query.Close;
+
+      UpdateMigrationVersion(Version);
+    end;
+
+    // Create Job Vacancy Table
+    if Version <= 5 then
+    begin
+      Version := Version + 1;
+
+      Query.SQL.Clear;
+      Query.SQL.Add('CREATE TABLE IF NOT EXISTS job_vacancy (           ');
+      Query.SQL.Add('   id INTEGER PRIMARY KEY AUTOINCREMENT,           ');
+      Query.SQL.Add('   title VARCHAR(100) NOT NULL,                    ');
+      Query.SQL.Add('   description TEXT,                               ');
+      Query.SQL.Add('   closed INT NOT NULL DEFAULT 0);                 ');
+      Query.ExecSQL;
+      Query.Close;
+
+      UpdateMigrationVersion(Version);
+    end;
+
+    // Create Job Vacancy Analysis Table
+    if Version <= 6 then
+    begin
+      Version := Version + 1;
+
+      Query.SQL.Clear;
+      Query.SQL.Add('CREATE TABLE IF NOT EXISTS job_vacancy_analysis (           ');
+      Query.SQL.Add('   id INTEGER PRIMARY KEY AUTOINCREMENT,                    ');
+      Query.SQL.Add('   job_vacancy_id INTEGER NOT NULL,                         ');
+      Query.SQL.Add('   file_id INTEGER NOT NULL,                                ');
+      Query.SQL.Add('   score DOUBLE,                                            ');
+      Query.SQL.Add('   details TEXT,                                            ');
+      Query.SQL.Add('   FOREIGN KEY (job_vacancy_id) REFERENCES job_vacancy(id), ');
+      Query.SQL.Add('   FOREIGN KEY (file_id) REFERENCES file(id));              ');
+      Query.ExecSQL;
+      Query.Close;
+
+      UpdateMigrationVersion(Version);
     end;
     
   finally
@@ -133,6 +236,26 @@ begin
   restRequest.Execute;
 
   Result := restResponse.StatusCode = 200;
+end;
+
+procedure TdtmMainDataModule.UpdateMigrationVersion(Version: Integer);
+var
+  Query: TFDQuery;
+begin
+  Query := TFDQuery.Create(nil);
+  try
+    Query.Connection := fdcDatabase;
+
+    Query.SQL.Clear;
+    Query.Params.Clear;
+    Query.SQL.Add('UPDATE parameter SET param_value = :param_value WHERE param_name = :param_name;');
+    Query.Params.ParamByName('param_name').Value := PARAM_VERSION_NAME;
+    Query.Params.ParamByName('param_value').Value := IntToStr(Version);
+    Query.ExecSQL;
+    Query.Close;
+  finally
+    Query.Free;
+  end;
 end;
 
 function TdtmMainDataModule.ValidatedMail(Hash: string): boolean;